+++ tests/test_multipart.py
@@ -1115,6 +1115,21 @@ class MultipartWriterTestCase(unittest.TestCase):
             writer.append_json({'baz': True})
         self.assertEqual(3, len(writer))
 
+    def test_append_int_not_allowed(self):
+        with self.assertRaises(TypeError):
+            with aiohttp.multipart.MultipartWriter(boundary=':') as writer:
+                writer.append(1)
+
+    def test_append_float_not_allowed(self):
+        with self.assertRaises(TypeError):
+            with aiohttp.multipart.MultipartWriter(boundary=':') as writer:
+                writer.append(1.1)
+
+    def test_append_none_not_allowed(self):
+        with self.assertRaises(TypeError):
+            with aiohttp.multipart.MultipartWriter(boundary=':') as writer:
+                writer.append(None)
+
 
 class ParseContentDispositionTestCase(unittest.TestCase):
     # http://greenbytes.de/tech/tc2231/